refactor: project management structure and add component architecture
