+++ src/app.ts
@@ -1,162 +1,11 @@
-//Project State Mangement
-class State<T> {
-    protected listeners: Listener<T>[] = [];
+import { autoBind } from './decorators.js';
+import { ProjectStatus } from './interfaces/project-status.js';
+import { Draggable, DragTarget } from './interfaces/draggable.js';
+import { Validatable, validate } from './helpers/validate.js';
+import { Project } from './interfaces/project.js';
+import { projectState } from './project-state.js';
+import { Component } from './component.js';
 
-    addListener(listenerFn: Listener<T>) {
-        this.listeners.push(listenerFn);
-    }
-}
-
-class ProjectState extends State<Project> {
-    private projects: Project[] = [];
-    private static instance: ProjectState;
-
-    constructor() {
-        super();
-    }
-
-    static getInstance() {
-        if (this.instance) {
-            return this.instance;
-        }
-        this.instance = new ProjectState();
-        return this.instance;
-    }
-
-    addProject(title: string, description: string, numOfPeople: number) { 
-        const newProject = new Project(
-            Math.random().toString(),
-            title,
-            description,
-            numOfPeople,
-            ProjectStatus.Active
-        );
-
-        this.projects.push(newProject);
-
-        this.updateListeners();
-    }
-
-    moveProject(projectId: string, newStatus: ProjectStatus) {
-        const project = this.projects.find(prj => prj.id === projectId);
-        if (project && project.status !== newStatus) {
-            project.status = newStatus;
-            this.updateListeners();
-        }
-    }
-
-    private updateListeners() {
-        for (const listenerFn of this.listeners) {
-            listenerFn(this.projects.slice());
-        }
-    }
-}
-const projectState = ProjectState.getInstance();   //Global Project State Object
-
-//Types
-interface Validatable {
-    value: string | number;
-    required?: boolean;
-    minLength?: number;
-    maxLength?: number;
-    min?: number;
-    max?: number;
-}
-
-class Project {
-    constructor(
-        public id: string, 
-        public title: string, 
-        public description: string, 
-        public people: number,
-        public status: ProjectStatus
-    ) {}
-}
-
-enum ProjectStatus { Active, Finished }
-
-type Listener<T> = (items: T[]) => void;
-
-//Drag And Drop Interfaces
-interface Draggable {
-    dragStartHandler(event: DragEvent): void;
-    dragEndHandler(event: DragEvent): void;
-}
-
-interface DragTarget {
-    dragOverHandler(event: DragEvent): void;
-    dropHandler(event: DragEvent): void;
-    dragLeaveHandler(event: DragEvent): void;
-}
-
-//Helpers
-function validate(validatableInput: Validatable) {
-    let isValid = true;
-    if (validatableInput.required) {
-        isValid = isValid && validatableInput.value.toString().trim().length !== 0;
-    }
-    if(validatableInput.minLength != null && typeof validatableInput.value === 'string') {
-        isValid = isValid && validatableInput.value.length >= validatableInput.minLength;
-    }
-    if(validatableInput.maxLength != null && typeof validatableInput.value === 'string') {
-        isValid = isValid && validatableInput.value.length <= validatableInput.maxLength;
-    }
-    if(validatableInput.min != null && typeof validatableInput.value === 'number') {
-        isValid = isValid && validatableInput.value >= validatableInput.min;
-    }
-    if(validatableInput.max != null && typeof validatableInput.value === 'number') {
-        isValid = isValid && validatableInput.value <= validatableInput.max;
-    }
-    return isValid;
-}
-
-//Decorators 
-function autoBind(
-    _: any, 
-    _2: string, 
-    descriptor: PropertyDescriptor
-) {
-    const originalMethod = descriptor.value;
-    const adjDescriptor: PropertyDescriptor = {
-        configurable: true,
-        get() {
-            const boundFn = originalMethod.bind(this);
-            return boundFn;
-        }
-    };
-    return adjDescriptor;
-}
-
-abstract class Component<T extends HTMLElement, U extends HTMLElement> {
-    templateElement: HTMLTemplateElement;
-    hostElement: T;
-    element: U;
-
-    constructor(
-        templateId: string, 
-        hostElementId: string,
-        insertAtStart: boolean, 
-        newElementId?: string
-    ) {
-        this.templateElement = document.getElementById(templateId)! as HTMLTemplateElement;
-        this.hostElement = document.getElementById(hostElementId)! as T;
-
-        const importedNode = document.importNode(this.templateElement.content, true);
-        this.element = importedNode.firstElementChild as U;
-        if (newElementId) {
-            this.element.id = newElementId;
-        }
-
-        this.attatch(insertAtStart);
-    }
-
-    private attatch(insertAtBeginning: boolean) {
-        this.hostElement.insertAdjacentElement(insertAtBeginning ? 'afterbegin':'beforeend', this.element);
-    }
-
-    abstract configure(): void;
-    abstract renderContent(): void;
-}
 
 class ProjectItem extends Component<HTMLUListElement, HTMLLIElement> implements Draggable {
     private project: Project;
@@ -345,6 +194,6 @@ class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
     }
 }
 
-const prjInput = new ProjectInput();
-const activePrjList = new ProjectList('active');
-const finishedPrjList = new ProjectList('finished');
+new ProjectInput();
+new ProjectList('active');
+new ProjectList('finished');
