+++ src/interfaces/project-status.ts
@@ -1,0 +1,1 @@
+export enum ProjectStatus { Active, Finished }
