+++ src/interfaces/project.ts
@@ -1,0 +1,11 @@
+import { ProjectStatus } from '../interfaces/project-status.js';
+
+export class Project {
+    constructor(
+        public id: string, 
+        public title: string, 
+        public description: string, 
+        public people: number,
+        public status: ProjectStatus
+    ) {}
+}
